# Desktop/aqarmatch-main/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AqarMatch API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: getListings
      tags: [listings]
      summary: Browse property listings
      parameters:
        - in: query
          name: deal_type
          schema:
            type: string
        - in: query
          name: wilaya
          schema:
            type: string
        - in: query
          name: municipality
          schema:
            type: string
        - in: query
          name: max_price
          schema:
            type: number
      responses:
        "200":
          description: List of listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
    post:
      operationId: createListing
      tags: [listings]
      summary: Create a new property listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingInput"
      responses:
        "201":
          description: Listing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "400":
          description: Validation error
...
  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get a single listing
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          description: Not found
  /listings/{id}/match:
    post:
      operationId: matchListing
      tags: [listings]
      summary: Check if buyer budget matches listing
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchRequest"
      responses:
        "200":
          description: Match result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResult"
        "404":
          description: Listing not found
  /listings/{id}/renew:
    post:
      operationId: renewListing
      tags: [listings]
      summary: Renew a listing for another 30 days
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenewRequest"
      responses:
        "200":
          description: Listing renewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "403":
          description: Phone number does not match listing owner
        "404":
          description: Listing not found
...
components:
  schemas:
...
    Listing:
      type: object
      required:
        - id
        - deal_type
        - wilaya
        - municipality
        - neighborhoods
        - asking_price
        - user_phone
        - created_at
        - expires_at
        - is_active
        - days_remaining
      properties:
        id:
          type: integer
        deal_type:
          type: string
          enum: [بيع, إيجار]
        wilaya:
          type: string
        municipality:
          type: string
        neighborhoods:
          type: array
          items:
            type: string
        asking_price:
          type: number
        user_phone:
          type: string
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        is_active:
          type: boolean
        days_remaining:
          type: integer
    ListingInput:
      type: object
      required:
        - deal_type
        - wilaya
        - municipality
        - neighborhoods
        - asking_price
        - floor_price
        - user_phone
      properties:
        deal_type:
          type: string
          enum: [بيع, إيجار]
        wilaya:
          type: string
        municipality:
          type: string
        neighborhoods:
          type: array
          items:
            type: string
        asking_price:
          type: number
        floor_price:
          type: number
        user_phone:
          type: string
    RenewRequest:
      type: object
      required:
        - seller_phone
      properties:
        seller_phone:
          type: string
    MatchRequest:
      type: object
      required:
        - budget
        - buyer_phone
      properties:
        budget:
          type: number
        buyer_phone:
          type: string
    MatchResult:
      type: object
      required:
        - matched
        - listing_id
      properties:
        matched:
          type: boolean
        listing_id:
          type: integer
        seller_phone:
          type: ["string", "null"]
        message:
          type: string
...