# Desktop/aqarmatch-main/lib/api-zod/src/generated/types/listing.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AqarMatch API
 * OpenAPI spec version: 0.1.0
 */
import type { ListingDealType } from './listingDealType';

export interface Listing {
  id: number;
  deal_type: ListingDealType;
  wilaya: string;
  municipality: string;
  neighborhoods: string[];
  asking_price: number;
  user_phone: string;
  created_at: Date;
  expires_at: Date;
  is_active: boolean;
  days_remaining: number;
}