# Desktop/aqarmatch-main/lib/api-zod/src/generated/types/listingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AqarMatch API
 * OpenAPI spec version: 0.1.0
 */
import type { ListingInputDealType } from './listingInputDealType';

export interface ListingInput {
  deal_type: ListingInputDealType;
  wilaya: string;
  municipality: string;
  neighborhoods: string[];
  asking_price: number;
  floor_price: number;
  user_phone: string;
}